02-agent-survey/01-langchain/requirements.txt
langchain==0.3.13
langchain-groq==0.2.3
langchain-community==0.3.13
python-dotenv==1.0.1
